requirements.txt
appdirs>=1.4.4
attrs>=21.4.0
CacheControl>=0.12.10
cattrs>=1.10.0
certifi>=2021.10.8
cffi>=1.15.0
charset-normalizer>=2.0.11
cryptography>=36.0.1
Deprecated>=1.2.13
et-xmlfile>=1.1.0
idna>=3.3
lockfile>=0.12.2
msgpack>=1.0.3
openpyxl>=3.0.9
packaging>=21.3
psycopg2>=2.9.3
pycparser>=2.21
PyMySQL>=1.0.2
pyparsing>=3.0.7
PyYAML>=6.0
redis>=4.1.2
requests>=2.27.1
requests-cache>=0.9.1
requests-futures>=1.0.0
six>=1.16.0
SQLAlchemy>=1.4.31
tqdm>=4.62.3
url-normalize>=1.4.3
urllib3>=1.26.8
wrapt>=1.13.3
